requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.22
pydantic==1.10.11
python-dotenv==1.0.0
psycopg2-binary==2.9.7
python-multipart==0.0.6
transformers>=4.30.0
torch>=2.0.0
spacy>=3.6.0
textblob>=0.17.1
nltk>=3.8.1
scikit-learn>=1.3.2
pandas>=2.0.3
numpy>=1.24.4
